Refactor test setup and utilities for improved testing infrastructure
