+++ package.json
@@ -10,7 +10,7 @@
     "lint": "bun eslint .",
     "preview": "bun vite preview",
     "test:ai": "tsx src/test-ai.ts",
-    "test": "vitest",
+    "test": "vitest run",
     "test:components": "vitest --run src/test/__tests__/components",
     "test:coverage": "vitest --coverage",
     "test:watch": "vitest --watch",
@@ -89,6 +89,7 @@
   "devDependencies": {
     "@eslint/js": "^9.32.0",
     "@tailwindcss/typography": "^0.5.16",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.3.1",
     "@types/bun": "latest",
     "@types/node": "^22.16.5",
