+++ src/test/vitest.d.ts
@@ -1,0 +1,2 @@
+/// <reference types="vitest/globals" />
+/// <reference types="@testing-library/jest-dom" />
